+++ server/src/config/passport.js
@@ -1,7 +1,6 @@
 const JwtStrategy = require('passport-jwt').Strategy;
 const ExtractJwt = require('passport-jwt').ExtractJwt;
 const Passport = require('passport');
-const config = require('./database');
 const User = require('../models/User');
 
 /**
@@ -11,7 +10,7 @@ const User = require('../models/User');
 module.exports = function (passport) {
   let options = {};
   options.jwtFromRequest = ExtractJwt.fromAuthHeaderWithScheme('jwt');
-  options.secretOrKey = config.secret;
+  options.secretOrKey = process.env.JWT_SECRET;
   passport.use(
     new JwtStrategy(options, (jwt_payload, done) => {
       User.findById(jwt_payload._id)
@@ -25,3 +24,5 @@ module.exports = function (passport) {
     })
   );
 };
+
+module.exports.authOptions = ['jwt', { session: false }];
